+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 
 [Unreleased]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.0.0rc2...HEAD
 
+### Changed
+
+- Fix differences of API between Vault KV secret v1 and v2 [#80][80]
+
+[80]: https://github.com/chaostoolkit/chaostoolkit-lib/issues/80
+
 ## [1.0.0rc2][] - 2019-01-28
 
 [1.0.0rc2]: https://github.com/chaostoolkit/chaostoolkit-lib/compare/1.0.0rc1...1.0.0rc2